requirements.txt
opencv-python>=4.5.5
torch>=1.9.0
torchvision>=0.10.0
tqdm
Pillow
clean-fid>=0.1.35